+++ Sources/EpoxyCollectionView/CollectionView/CollectionView.swift
@@ -445,6 +445,7 @@ open class CollectionView: UICollectionView {
     }
 
     cell.itemPath = itemPath
+    cell.forcesEarlySwiftUIRendering = configuration.forcesEarlySwiftUIRendering
 
     let metadata = ItemCellMetadata(
       traitCollection: traitCollection,
@@ -469,6 +470,7 @@ open class CollectionView: UICollectionView {
     animated: Bool)
   {
     supplementaryView.itemPath = itemPath
+    supplementaryView.forcesEarlySwiftUIRendering = configuration.forcesEarlySwiftUIRendering
     model.configure(
       reusableView: supplementaryView,
       traitCollection: traitCollection,