+++ Sources/EpoxyCollectionView/CollectionView/ReusableViews/CollectionViewReusableView.swift
@@ -1,10 +1,11 @@
 //  Created by Laura Skelton on 9/8/17.
 //  Copyright © 2017 Airbnb. All rights reserved.
 
+import EpoxyCore
 import UIKit
 
 /// An internal collection reusable view class for use in a `CollectionView`.
-public final class CollectionViewReusableView: UICollectionReusableView {
+public final class CollectionViewReusableView: UICollectionReusableView, SwiftUIRenderingConfigurable {
 
   // MARK: Lifecycle
 
@@ -21,6 +22,13 @@ public final class CollectionViewReusableView: UICollectionReusableView {
   // MARK: Public
 
   public private(set) var view: UIView?
+
+  /// See `CollectionViewConfiguration.forcesEarlySwiftUIRendering` for an explanation of this behavior.
+  public var forcesEarlySwiftUIRendering = false {
+    didSet {
+      updateForcesEarlySwiftUIRendering()
+    }
+  }
 
   /// Pass a view for this view's element kind and reuseID that the view will pin to its edges.
   public func setViewIfNeeded(view: UIView) {
@@ -39,6 +47,8 @@ public final class CollectionViewReusableView: UICollectionReusableView {
       view.bottomAnchor.constraint(equalTo: bottomAnchor),
     ])
     self.view = view
+
+    updateForcesEarlySwiftUIRendering()
   }
 
   override public func preferredLayoutAttributesFitting(
@@ -75,5 +85,13 @@ public final class CollectionViewReusableView: UICollectionReusableView {
   /// view provides view display callbacks, if it is mid update, we need this to see if the view came from pre-update data or
   /// post-update data.
   var itemPath: SupplementaryItemPath?
+
+  // MARK: Private
+
+  private func updateForcesEarlySwiftUIRendering() {
+    if let configurableView = view as? SwiftUIRenderingConfigurable {
+      configurableView.forcesEarlySwiftUIRendering = forcesEarlySwiftUIRendering
+    }
+  }
 
 }