+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   batch update.
 - Added caching for `visibilityMetadata` calculations.
 - Fixed an issue that could cause SwiftUI views to be incorrectly sized in a collection view.
+- Added `forcesEarlySwiftUIRendering` flag to `CollectionViewConfiguration` to test a SwiftUI layout
+  approach to resolve an issue that could cause collection view cells to layout with 
+  unexpected dimensions
 
 ## [0.10.0](https://github.com/airbnb/epoxy-ios/compare/0.9.0...0.10.0) - 2023-06-29